rest-api response creator as a service
