+++ restApi/src/main/scala/de/htwberlin/f4/wikiplag/rest/models/Text.scala
@@ -1,6 +1,8 @@
 package de.htwberlin.f4.wikiplag.rest.models
 
-/** The entity accepted by the analyse endpoint of the Wikiplag servlet. */
+/**
+  * The entity accepted by the analyse endpoint of the Wikiplag servlet.
+  */
 case class Text(text: String) {
 }
